+++ Neolution.DotNet.Console.UnitTests/CollectionDefinitions/ProductionEnvironmentTestCollection.cs
@@ -1,11 +1,12 @@
-﻿namespace Neolution.DotNet.Console.UnitTests.Common.CollectionDefinitions
+﻿namespace Neolution.DotNet.Console.UnitTests.CollectionDefinitions
 {
+    using Neolution.DotNet.Console.UnitTests.Setup;
     using Xunit;
 
     /// <summary>
     /// Collection definition for tests that should run in the production environment.
     /// </summary>
-    /// <seealso cref="Xunit.ICollectionFixture{ProductionEnvironmentTestSetup}" />
+    /// <seealso cref="ICollectionFixture{ProductionEnvironmentTestSetup}" />
     [CollectionDefinition("Production Environment Tests")]
     public class ProductionEnvironmentTestCollection : ICollectionFixture<ProductionEnvironmentTestSetup>
     {