+++ Neolution.DotNet.Console.UnitTests/DotNetConsoleBuilderTests.cs
@@ -1,17 +1,16 @@
-﻿namespace Neolution.DotNet.Console.UnitTests.GrammarTests
+﻿namespace Neolution.DotNet.Console.UnitTests
 {
     using System;
     using System.Collections.Generic;
     using System.Reflection;
-    using Neolution.DotNet.Console.UnitTests.GrammarTests.Fakes;
+    using Neolution.DotNet.Console.UnitTests.Fakes;
     using Shouldly;
     using Xunit;
 
     /// <summary>
     /// Command line arguments grammar tests
     /// </summary>
-    [Collection("Production Environment Tests")]
-    public class StrictVerbMatchingGrammarTests
+    public class DotNetConsoleBuilderTests
     {
         /// <summary>
         /// Given a mistyped verb, when a default verb is defined, then should throw on console building.
@@ -29,25 +28,27 @@
         }
 
         /// <summary>
-        /// Given a mistyped verb, when no default verb is defined, then should not throw on console building.
+        /// Given a valid argument, when default verb is defined, then should not throw on building.
         /// </summary>
-        [Fact]
-        public void GivenMistypedVerb_WhenNoDefaultVerbIsDefined_ThenShouldNotThrowOnBuilding()
+        /// <param name="args">The arguments.</param>
+        [Theory]
+        [InlineData("")]
+        [InlineData("echo")]
+        [InlineData("--silent")]
+        [InlineData("-s")]
+        public void GivenValidArgument_WhenDefaultVerbIsDefined_ThenShouldNotThrowOnBuilding(string args)
         {
             // Arrange
-            const string args = "eho"; // mistyped verb
-
             var servicesAssembly = Assembly.GetAssembly(typeof(EchoCommand))!;
-            var verbTypes = new List<Type> { typeof(EchoOptions) }.ToArray();
 
             // Act
 
             // Assert
-            Should.NotThrow(() => DotNetConsole.CreateBuilderWithReference(servicesAssembly, verbTypes, args.Split(" ")));
+            Should.NotThrow(() => DotNetConsole.CreateBuilderWithReference(servicesAssembly, args.Split(" ")));
         }
 
         /// <summary>
-        /// Givens a reserved argument name, when no default verb is defined, then should not throw on building.
+        /// Given a reserved argument name, when default verb is defined, then should not throw on building.
         /// </summary>
         /// <param name="args">The arguments.</param>
         [Theory]
@@ -56,7 +57,27 @@
         [InlineData("--help")]
         [InlineData("--version")]
         [InlineData("help echo")]
-        public void GivenReservedArgumentName_WhenNoDefaultVerbIsDefined_ThenShouldNotThrowOnBuilding(string args)
+        public void GivenReservedArgumentName_WhenDefaultVerbIsDefined_ThenShouldNotThrowOnBuilding(string args)
+        {
+            // Arrange
+            var servicesAssembly = Assembly.GetAssembly(typeof(EchoCommand))!;
+
+            // Act
+
+            // Assert
+            Should.NotThrow(() => DotNetConsole.CreateBuilderWithReference(servicesAssembly, args.Split(" ")));
+        }
+
+        /// <summary>
+        /// Given invalid arguments, when no default verb is defined, then should not throw on console building.
+        /// </summary>
+        /// <param name="args">The arguments.</param>
+        [Theory]
+        [InlineData("")]
+        [InlineData("verb-that-does-not-exist")]
+        [InlineData("--option-that-does-not-exist")]
+        [InlineData("-o")]
+        public void GivenInvalidArguments_WhenNoDefaultVerbIsDefined_ThenShouldNotThrowOnBuilding(string args)
         {
             // Arrange
             var servicesAssembly = Assembly.GetAssembly(typeof(EchoCommand))!;